pyproject.toml
[project]
name = "groq-chatbot-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
]
